move Transitivity and Annotation to concept
