+++ src/lib.rs
@@ -28,10 +28,7 @@ mod connection;
 mod database;
 
 pub use self::{
-    common::{
-        error, Annotation, Credential, Error, Options, Result, SchemaException, SessionType, TransactionType,
-        Transitivity,
-    },
+    common::{error, Credential, Error, Options, Result, SchemaException, SessionType, TransactionType},
     connection::Connection,
     database::{Database, DatabaseManager, Session, Transaction},
 };
